Make leaf node comparisons more efficient
Instead of allocating boolean vectors representing matrices of a permuted
graphs and comparing them, comparisons of graph matrices under permutations
should be done on the fly. This can be done since the comparison is carried out
in a lexicographic fashion which corresponds exactly to the order in which the
matrices (or bool-vectors) are generated. This removes many unnecessary memory
allocations and in most cases terminates the comparison early.