Fix and/or clean the search function and separate it from the main(), add --help option
Max path and aut path conditions were incorrect in some edge cases e.g. an aut_path could be
detected if the node invariant at this level is equal the node invariant of the fst_node at this level,
even if they are different on the previous level.

Max node and aut node are now represented by the node struct. Max lca level calculation no longer
requires the previous value to be passed down the tree. At all times the value is equal the lca level
of the current node and the maximal path node.

Automorphism detection is handled much more cleanly. If an automorphism is detected using the first
terminal node, it is possible to ignore the rest of the subtree and we backjump to the longest common
ancestor of the current and first nodes. Furthermore, we may ignore the maximal path node at this point
since it is guaranteed that the first and maximal path nodes have different matrix representations (except
for the case when they are the same node). On the other hand, if no automorphism is detected with the
first node, the automorphism check and possibly backjump (to max lca) is performed using the maximal
path node.