Fix orbit partition calculation
Calculating the set of minimal cell representatives by only intersecting does
not produce the mcr set of the orbit partition. Now, the partition is recalculated
for every newly found automorphism using breadth first search.

Due to this change, it is now possible for the automorphism detected via the maximal
node to prune the tree at the first node lowest common ancestor level. That case is
now being handled.