Fix segfault caused by incorrect lca values
All levels, including lca variables, denote the length of the stabilized
sequence of vertices. This means that in order to find the vertex at level L
one should look up the stabilized element at index L, not index L+1 as was
implemented. The indexing error caused segfaults and possibly some silent
errors in the execution and the results of the algorithm.