Replace old multiset hash function
The modpow rolling hash function seems to have induced too many collisions
and on top of that was not efficient to calculate. It is now replaced with
the sum of hashes of single elements which seems to fix a lot of the problems
with search tree sizes and the inconsistency thereof.